update file name and comptroller address, and fixed lint issues
