+++ src/apps/iron-bank/ethereum/iron-bank.tvl-fetcher.ts
@@ -1,11 +1,11 @@
 import { Inject } from '@nestjs/common';
 
 import { Register } from '~app-toolkit/decorators';
+import { CompoundTvlHelper } from '~apps/compound/helper/compound.tvl-helper';
 import { TvlFetcher } from '~stats/tvl/tvl-fetcher.interface';
 import { Network } from '~types/network.interface';
 
 import { IRON_BANK_DEFINITION } from '../iron-bank.definition';
-import { CompoundTvlHelper } from '~apps/compound/helper/compound.tvl-helper';
 
 const appId = IRON_BANK_DEFINITION.id;
 const network = Network.ETHEREUM_MAINNET;
