+++ src/apps/iron-bank/fantom/iron-bank.balance-fetcher.ts
@@ -1,16 +1,15 @@
-import { Inject, forwardRef } from '@nestjs/common';
+import { Inject } from '@nestjs/common';
 
 import { Register } from '~app-toolkit/decorators';
 import { presentBalanceFetcherResponse } from '~app-toolkit/helpers/presentation/balance-fetcher-response.present';
+import { CompoundBorrowBalanceHelper } from '~apps/compound/helper/compound.borrow.balance-helper';
+import { CompoundLendingMetaHelper } from '~apps/compound/helper/compound.lending.meta-helper';
+import { CompoundSupplyBalanceHelper } from '~apps/compound/helper/compound.supply.balance-helper';
 import { BalanceFetcher } from '~balance/balance-fetcher.interface';
 import { Network } from '~types/network.interface';
 
-import { IRON_BANK_DEFINITION } from '../iron-bank.definition';
 import { IronBankContractFactory } from '../contracts';
-
-import { CompoundLendingMetaHelper } from '~apps/compound/helper/compound.lending.meta-helper';
-import { CompoundSupplyBalanceHelper } from '~apps/compound/helper/compound.supply.balance-helper';
-import { CompoundBorrowBalanceHelper } from '~apps/compound/helper/compound.borrow.balance-helper';
+import { IRON_BANK_DEFINITION } from '../iron-bank.definition';
 
 const appId = IRON_BANK_DEFINITION.id;
 const network = Network.FANTOM_OPERA_MAINNET;
@@ -52,12 +51,12 @@ export class FantomIronBankBalanceFetcher implements BalanceFetcher {
   async getBalances(address: string) {
     const [supplyBalances, borrowBalances] = await Promise.all([
       this.getSupplyBalances(address),
-      this.getBorrowBalances(address)
+      this.getBorrowBalances(address),
     ]);
 
     const meta = this.compoundLendingMetaHelper.getMeta({ balances: [...supplyBalances, ...borrowBalances] });
     const lendingProduct = { label: 'Lending', assets: [...supplyBalances, ...borrowBalances], meta };
 
     return presentBalanceFetcherResponse([lendingProduct]);
-}
+  }
 }
