+++ src/apps/iron-bank/fantom/iron-bank.borrow.contract-position-fetcher.ts
@@ -1,19 +1,19 @@
 import { Inject } from '@nestjs/common';
 
 import { Register } from '~app-toolkit/decorators';
+import { CompoundBorrowContractPositionHelper } from '~apps/compound/helper/compound.borrow.contract-position-helper';
 import { PositionFetcher } from '~position/position-fetcher.interface';
 import { ContractPosition } from '~position/position.interface';
 import { Network } from '~types/network.interface';
 
 import { IRON_BANK_DEFINITION } from '../iron-bank.definition';
-import { CompoundBorrowContractPositionHelper } from '~apps/compound/helper/compound.borrow.contract-position-helper';
 
 const appId = IRON_BANK_DEFINITION.id;
 const groupId = IRON_BANK_DEFINITION.groups.borrow.id;
-const network = Network.AVALANCHE_MAINNET;
+const network = Network.FANTOM_OPERA_MAINNET;
 
 @Register.ContractPositionFetcher({ appId, groupId, network })
-export class AvalancheIronBankBorrowContractPositionFetcher implements PositionFetcher<ContractPosition> {
+export class FantomIronBankBorrowContractPositionFetcher implements PositionFetcher<ContractPosition> {
   constructor(
     @Inject(CompoundBorrowContractPositionHelper)
     private readonly compoundBorrowContractPositionHelper: CompoundBorrowContractPositionHelper,
